Move source files from utility to initialization
